Moved interrupt sequence to a cli command
