+++ run_qemu.sh
@@ -1,5 +1,9 @@
 #! /usr/bin/env sh
 
+if [ -z "$QEMU_BIN" ]; then
+  QEMU_BIN=qemu-system-aarch64
+fi
+
 # If KERNEL_DEBUG is defined, set QEMU_ARGS to -s -S and set RELEASE_DIR to debug
 if [ -n "$KERNEL_DEBUG" ]; then
   QEMU_ARGS="-s -S"
@@ -16,7 +20,7 @@ fi
 # Run QEMU for aarch64 with the gicv3 and no security extensions
 MACHINE_ARGS=gic-version=3,secure=off
 
-qemu-system-aarch64       \
+$QEMU_BIN                 \
   -machine virt           \
   -cpu cortex-a53         \
   -nographic              \
